requirements.txt
opencv-python-headless
numpy
pillow
streamlit
pandas